move `SoftDeletedMetadataTable` from `media-api` to `common-lib` along with accompanying config property move from `MediaApiConfig` to `CommonConfig`
